requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.2
djangorestframework==3.16.1
drf-yasg==1.21.11
django-cors-headers==4.9.0
django-mongodb-backend==6.0.2
dnspython==2.8.0
idna==3.11
pymongo==4.16.0
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
